TAC-278 - changed taxRate percentages retrieved from the backend

+++ module/Products/src/Products/Product/TaxRate/Service.php
@@ -66,7 +66,7 @@ class Service
         foreach ($rates->getAll() as $rateId => $rate) {
             $ratesOptions[$rateId] = [
                 'name' => $rate->getName(),
-                'rate' => bcmul($rate->getCurrent(), 100, 2)
+                'rate' => (float) bcmul($rate->getCurrent(), 100, 2)
             ];
         }
         return $ratesOptions;
